# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter API",
    "description": "\nA clean, read-only REST API over the **Urja Meter Ops** portal (smart-meter data for a\nJaipur distribution utility). The portal has no API of its own. This service logs in as a\nnormal user, reads the portal's internal endpoints (see `PROTOCOL.md` in the repository)\nand serves the data as typed JSON.\n\n* **Meters**: nameplate, status, location, network position; filter by any attribute,\n  network node or distance.\n* **Readings / consumption**: register readings normalised to numbers and ISO timestamps\n  (IST), with per-interval and per-day consumption derived from register deltas.\n* **Network**: the zone \u2192 circle \u2192 division \u2192 subdivision \u2192 substation \u2192 feeder \u2192\n  transformer hierarchy, reconstructed and repaired from per-meter data.\n* **Data quality**: every value we repaired or could not reconcile, with the portal's\n  original value.\n\nReference data is served from a local index synced from the portal's bulk export;\nreadings are fetched on demand and cached. `GET /v1/status` shows how fresh everything is.\nErrors use `application/problem+json` (RFC 9457) with a stable `code` (listed under the\n`Problem` schema). Unknown query parameters are rejected rather than ignored.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Sync freshness and portal client health",
        "operationId": "status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/v1/meters": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "List and filter meters",
        "operationId": "list_meters",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring of meter id or serial number.",
              "title": "Q"
            },
            "description": "Case-insensitive substring of meter id or serial number."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeterStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable, e.g. `status=faulty&status=installed`.",
              "title": "Status"
            },
            "description": "Repeatable, e.g. `status=faulty&status=installed`."
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable, case-insensitive, e.g. `make=HPL`.",
              "title": "Make"
            },
            "description": "Repeatable, case-insensitive, e.g. `make=HPL`."
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Phase"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Single- or three-phase meters only.",
              "title": "Phase"
            },
            "description": "Single- or three-phase meters only."
          },
          {
            "name": "installation_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InstallationType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Whole-current or CT-operated meters only.",
              "title": "Installation Type"
            },
            "description": "Whole-current or CT-operated meters only."
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Zone code, e.g. `Z-02`.",
              "title": "Zone"
            },
            "description": "Zone code, e.g. `Z-02`."
          },
          {
            "name": "circle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Circle code.",
              "title": "Circle"
            },
            "description": "Circle code."
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Division code.",
              "title": "Division"
            },
            "description": "Division code."
          },
          {
            "name": "subdivision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Subdivision code.",
              "title": "Subdivision"
            },
            "description": "Subdivision code."
          },
          {
            "name": "substation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substation code.",
              "title": "Substation"
            },
            "description": "Substation code."
          },
          {
            "name": "feeder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Feeder code, e.g. `F-001` (the meters' `feeder_code`).",
              "title": "Feeder"
            },
            "description": "Feeder code, e.g. `F-001` (the meters' `feeder_code`)."
          },
          {
            "name": "transformer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Distribution transformer code, e.g. `DT-001` (the meters' `transformer_code`).",
              "title": "Transformer"
            },
            "description": "Distribution transformer code, e.g. `DT-001` (the meters' `transformer_code`)."
          },
          {
            "name": "has_issues",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only meters with (or without) data-quality issues.",
              "title": "Has Issues"
            },
            "description": "Only meters with (or without) data-quality issues."
          },
          {
            "name": "interval_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reading interval: 30 (half-hourly) or 1440 (daily). Only meters whose readings are cached can match (`/v1/status` shows how many are).",
              "title": "Interval Minutes"
            },
            "description": "Reading interval: 30 (half-hourly) or 1440 (daily). Only meters whose readings are cached can match (`/v1/status` shows how many are)."
          },
          {
            "name": "near",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`lat,lng`: return meters within `radius_km`, nearest first, with `distance_km`.",
              "examples": [
                "26.9124,75.7873"
              ],
              "title": "Near"
            },
            "description": "`lat,lng`: return meters within `radius_km`, nearest first, with `distance_km`."
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 50,
              "exclusiveMinimum": 0,
              "description": "Search radius around `near`, in km.",
              "default": 1.0,
              "title": "Radius Km"
            },
            "description": "Search radius around `near`, in km."
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`min_lng,min_lat,max_lng,max_lat` (GeoJSON order). With `near=`, a meter must be in both.",
              "examples": [
                "75.78,26.90,75.82,26.94"
              ],
              "title": "Bbox"
            },
            "description": "`min_lng,min_lat,max_lng,max_lat` (GeoJSON order). With `near=`, a meter must be in both."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^-?(meter_id|serial_number|make|status|phase|installation_type|transformer|feeder|issues|interval|distance|transformer_code|feeder_code|data_issue_count|reading_interval_minutes|distance_km)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort key, prefixed with `-` for descending: `meter_id`, `serial_number`, `make`, `status`, `phase`, `installation_type`, `transformer`, `feeder`, `issues`, `interval`, `distance` (`distance` needs `near=`). The response's field names (`transformer_code`, `data_issue_count`, ...) work too. Defaults to `distance` with `near=`, else `meter_id`; ties are broken by `meter_id`.",
              "examples": [
                "-issues"
              ],
              "title": "Sort"
            },
            "description": "Sort key, prefixed with `-` for descending: `meter_id`, `serial_number`, `make`, `status`, `phase`, `installation_type`, `transformer`, `feeder`, `issues`, `interval`, `distance` (`distance` needs `near=`). The response's field names (`transformer_code`, `data_issue_count`, ...) work too. Defaults to `distance` with `near=`, else `meter_id`; ties are broken by `meter_id`."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterListItem_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Get one meter",
        "operationId": "get_meter",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The meter's id as the portal shows it (any case accepted).",
              "examples": [
                "J100000"
              ],
              "title": "Meter Id"
            },
            "description": "The meter's id as the portal shows it (any case accepted)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/meters/{meter_id}/anomalies": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Anomalies detected in one meter's readings",
        "description": "Every anomaly rule (listed under `/v1/insights/anomalies`) run over the meter's\nwhole cached series, most severe first.",
        "operationId": "get_meter_anomalies",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The meter's id as the portal shows it (any case accepted).",
              "examples": [
                "J100000"
              ],
              "title": "Meter Id"
            },
            "description": "The meter's id as the portal shows it (any case accepted)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterAnomalies"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InstallationType": {
        "type": "string",
        "enum": [
          "whole_current",
          "ct_operated",
          "unknown"
        ],
        "title": "InstallationType"
      },
...
      "MeterStatus": {
        "type": "string",
        "enum": [
          "installed",
          "faulty",
          "decommissioned",
          "unknown"
        ],
        "title": "MeterStatus"
      },
...
      "Page_MeterListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeterListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of items matching the query (ignoring limit/offset)."
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MeterListItem]"
      },
...
      "Phase": {
        "type": "string",
        "enum": [
          "single",
          "three",
          "unknown"
        ],
        "title": "Phase"
      },
...
      "Status": {
        "properties": {
          "index_ready": {
            "type": "boolean",
            "title": "Index Ready",
            "description": "True once reference data has been synced at least once."
          },
          "reference_data": {
            "$ref": "#/components/schemas/ReferenceDataStatus"
          },
          "readings_cache": {
            "$ref": "#/components/schemas/ReadingsCacheStatus"
          },
          "portal": {
            "$ref": "#/components/schemas/PortalClientStatus"
          }
        },
        "type": "object",
        "required": [
          "index_ready",
          "reference_data",
          "readings_cache",
          "portal"
        ],
        "title": "Status"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Required only if the server is configured with URJA_API_KEY.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}